Changed the way microauth is started.
