+++ microauth/__init__.py
@@ -1,14 +1,21 @@
 import os
 import bcrypt
+import optparse
 from flask.ext import restful
+from pkgutil import extend_path
 from flask.ext.sqlalchemy import SQLAlchemy
 from flask import Flask, send_from_directory
 from sqlalchemy.engine.reflection import Inspector
 
+__path__ = extend_path(__path__, __name__)
+__all__ = ["run", "repl", "resources"]
+
 app = Flask("microauth")
 app.config.from_object("microauth.config")
-db = SQLAlchemy(app)
+
 api = restful.Api(app, prefix='/v1')
+
+db = SQLAlchemy(app)
 
 from microauth.resources import api_key
 from microauth.resources import users
@@ -16,9 +23,11 @@ from microauth.resources import roles
 from microauth.resources import privs
 from microauth.resources import models
 
+
 inspector = Inspector.from_engine(db.engine)
 tables = [table_name for table_name in inspector.get_table_names()]
-if 'users' not in tables:
+
+if __name__ == "__main__" and 'users' not in tables:
 	db.create_all()
 	master = models.APIKey(name = app.config['MASTER_KEY_NAME'])
 	if app.config['MASTER_KEY']: master.key = app.config['MASTER_KEY']
@@ -29,6 +38,8 @@ if 'users' not in tables:
 	master.global_del = True
 	db.session.add(master)
 	db.session.commit()
+
+
 
 api.add_resource(api_key.KeyCollection, "/keys")
 api.add_resource(api_key.KeyResource,   "/keys/<string:name>")
@@ -46,7 +57,6 @@ api.add_resource(roles.RevokePrivs,     "/roles/<string:name>/revoke")
 api.add_resource(privs.PrivCollection,  "/privs")
 api.add_resource(privs.PrivResource,    "/privs/<string:name>")
 
-
 @app.route('/')
 def docs_index():
 	f = open('docs/_build/html/index.html')
@@ -57,3 +67,5 @@ def docs_index():
 @app.route('/<path:path>')
 def docs(path):
 	return send_from_directory('docs/_build/html', path)
+
+
